check null children while deteleTxDesc

+++ x10.runtime/src-x10/x10/util/resilient/localstore/tx/commit/CommitHandler.x10
@@ -71,7 +71,7 @@ public abstract class CommitHandler {
                 }
             }
         }finally {
-            if (deleteTxDesc) {
+            if (deleteTxDesc && childrenVirtual != null) {
                 plh().txDescManager.delete(id, true);
             }
         }
